% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanPlot.R
\name{buildManhattanPlot}
\alias{buildManhattanPlot}
\title{Build a Manhattan plot of biomarkers}
\usage{
buildManhattanPlot(
  biomarkerDf = NULL,
  chromosomeDf = NULL,
  tissue = "",
  compound = "",
  mDataType = "",
  pValCutoff = 0.05,
  relativeGenomeCoords = TRUE,
  xLabel = NULL,
  yLabel = NULL,
  title = NULL
)
}
\arguments{
\item{biomarkerDf}{A data.frame of drug sensitivity biomarkers that you want
to plot, including the compound, tissue, and molecular data type used in the
experiment, the genomic coordinates of the genes which were tested on, and
the resulting p-value/fdr}

\item{chromosomeDf}{A data.frame of the reference genome, including the
lengths of all the chromosomes. The row order should match the chromosome
order in the reference genome. If there is an "all" row, it should be first.}

\item{tissue}{A string representing the tissue for which you want to plot
biomarkers}

\item{compound}{A string representing the compound for which you want to plot
biomarkers}

\item{mDataType}{A string representing the molecular data type for which you
want to plot biomarkers}

\item{pValCutoff}{(optional) A decimal number indicating the cutoff value for
significant observations; any results with a higher p-value will be grayed
out on the plot. Default value is 0.05.}

\item{relativeGenomeCoords}{(optional) A boolean that indicates whether the
genomic coordinates given in biomarkerDf are relative to the chromosome
(TRUE) or relative to the entire genome (absolute, FALSE). Default is TRUE.}

\item{xLabel}{(optional) The label for the x-axis. Defaults to "Estimate".}

\item{yLabel}{(optional) The label for the y-axis. Defaults to
"Log10 P-Value".}

\item{title}{(optional) The title for the plot. Defaults to "Drug
sensitivity in <tissueName> tissue in response to <compoundName>}
}
\value{
A ggplot2 plot object mapping the biomarkers of the experiment
(x-axis = genome position; y-axis = -log10(p-value or fdr))
TODO: update!
}
\description{
Build a Manhattan plot of biomarkers
}
\examples{
data(Biomarkers)
data(GRCh38.p13.Assembly)
buildManhattanPlot(Biomarkers, GRCh38.p13.Assembly, "Lung", "Trametinib",
                   "rna", pValCutoff=0.01, relativeGenomeCoords=TRUE)

}
