% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{buildVolcanoPlot}
\alias{buildVolcanoPlot}
\title{Build a volcano plot of biomarkers}
\usage{
buildVolcanoPlot(
  biomarkerDf,
  tissue = "",
  compound = "",
  mDataType = "",
  pValCutoff = 0.05,
  xLabel = NULL,
  yLabel = NULL,
  title = NULL
)
}
\arguments{
\item{biomarkerDf}{A data.frame of drug sensitivity biomarkers that you want
to plot, including the compound, tissue, and molecular data type used in the
experiment, the genomic coordinates of the genes which were tested on, and
the resulting p-value/fdr and estimate}

\item{tissue}{A string representing the tissue for which you want to plot
biomarkers}

\item{compound}{A string representing the compound for which you want to plot
biomarkers}

\item{mDataType}{A string representing the molecular data type for which you
want to plot biomarkers}

\item{pValCutoff}{A decimal number indicating the cutoff value for
significant observations; any results with a higher p-value will be grayed
out on the plot. Default value is 0.05.}

\item{xLabel}{(optional) The label for the x-axis. Defaults to "Estimate".}

\item{yLabel}{(optional) The label for the y-axis. Defaults to
"Log10 P-Value".}

\item{title}{(optional) The title for the plot. Defaults to "Drug
sensitivity in <tissueName> tissue in response to <compoundName>}
}
\value{
A ggplot2 plot object mapping the biomarkers of the experiment
(x-axis = estimate; y-axis = -log10(p-value))
#TODO: update
}
\description{
Note that this volcano plot is NOT plotting results of differential gene
expression analysis but is plotting the drug sensitivity signature
}
\examples{
data(Biomarkers)
buildVolcanoPlot(Biomarkers, "Lung", "Trametinib", "rna", 0.005)

}
