% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfallPlot.R
\name{buildWaterfallPlot}
\alias{buildWaterfallPlot}
\title{Build a waterfall plot of ???}
\usage{
buildWaterfallPlot(pdxDf)
}
\arguments{
\item{pdxDf}{A data.frame of ??}
}
\value{
A ggplot2 plot object mapping ???
}
\description{
More description???
}
\examples{
data(PDXE)
df <- processXevaSet(PDXE, "BRCA", "paclitaxel", "RNASeq")
buildWaterfallPlot(df)

}
