% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareExpressionToReference.R
\name{computeBiomarkerPercentile}
\alias{computeBiomarkerPercentile}
\title{Compute the percentile of a biomarker relative to some reference population.}
\usage{
computeBiomarkerPercentile(feature, value, reference)
}
\arguments{
\item{feature}{`character(1)` Feature name of the biomarkers of interest. Must
be a valid row.name in the reference `data.frame`.}

\item{value}{`numeric(1)` Value of the feature, must be in the same units
as the data in `reference`.}

\item{reference}{`data.frame` or `matrix` where rows are features, columns
are samples, and values are a measurement on each feature for each sample.
Rownames must match the specified `feature` value.}
}
\value{
`numeric(1)` Percentile of `value` relative to the distrubtion of
  `feature` in the `reference` population. Computed using `stats::ecdf`.
}
\description{
Compute the percentile of a biomarker relative to some reference population.
}
\seealso{
`stats::ecdf`
}
