% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnrichment.R
\name{computeGeneSetSimilarity}
\alias{computeGeneSetSimilarity}
\title{Compute the similarity between gene sets}
\usage{
computeGeneSetSimilarity(geneSets, similarityMetric = "overlap")
}
\arguments{
\item{geneSets}{A data.table or data.frame with columns "gs_name" (gene set
name) and "ensembl_gene" (ENSEMBL gene ID) which lists the genes in each
gene set}

\item{similarityMetric}{(optional) The type of similarity metric to compute.
Currently, the only option is "overlap", which calculates the proportion of
intersecting genes to total genes between each pair of gene sets.}
}
\value{
A data.frame that lists the similarity score between each pair of
gene sets. There will be three columns: "gs1" (gene set 1), "gs2" (gene set
2), and "similarity".
}
\description{
Compute the similarity between gene sets
}
\examples{
geneSetNames <- queryGene("ENSG00000000971", "GO:BP")
geneSets <- expandGeneSets(geneSetNames, "GO:BP")
computeGeneSetSimilarity(geneSets)

}
