% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlotPercentile.R
\name{densityPlotBiomarkerPercentile}
\alias{densityPlotBiomarkerPercentile}
\title{Compute the percentile of a biomarker relative to some reference population.}
\usage{
densityPlotBiomarkerPercentile(feature, value, reference)
}
\arguments{
\item{feature}{`character(1)` Feature name of the biomarkers of interest. Must
be a valid row.name in the reference `data.frame`.}

\item{value}{`numeric(1)` Value of the feature, must be in the same units
as the data in `reference`.}

\item{reference}{`data.frame` or `matrix` where rows are features, columns
are samples, and values are a measurement on each feature for each sample.
Rownames must match the specified `feature` value.}
}
\value{
`ggplot` Plot object, showing the percentile of `value` relative to
  the distribution of `feature` in `reference`.
}
\description{
Compute the percentile of a biomarker relative to some reference population.
}
