% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnrichment.R
\name{expandGeneSets}
\alias{expandGeneSets}
\title{Expand gene sets}
\usage{
expandGeneSets(geneSetNames, geneSetType = NULL)
}
\arguments{
\item{geneSetNames}{A character vector containing gene set names from MSigDb}

\item{geneSetType}{(optional) The types of gene sets in geneSetNames; see
MSigDb for possible subcategories (use msigdbr::msigdbr_collections()). The
geneSetNames alone are enough but providing the geneSetType may speed up
computation because less data will be retrieved from MSigDb.}
}
\value{
A data.table mapping all gene sets to their component genes (2
columns: "gs_name", "ensembl_gene")
}
\description{
Get all genes that are in the MSigDb gene sets corresponding to geneSetNames.
}
\examples{
geneSetNames <- queryGene("ENSG00000000971", "GO:BP")
expandGeneSets(geneSetNames, "GO:BP")

}
