% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchBiomarkersPharmacoDB.R
\name{fetchPharmacoDbBiomarkers}
\alias{fetchPharmacoDbBiomarkers}
\title{Retrieve statistically significant biomarkers per compound per tissue from
  PharmacoDB.ca}
\usage{
fetchPharmacoDbBiomarkers(
  mysql = file.path("~", ".mysql"),
  url = "PHARMACODB",
  port = "PHARMACODB_PORT",
  user = "MYSQL_USER",
  password = "MYSQL_PW"
)
}
\arguments{
\item{url}{`character(1)` Name of enviromental variable with database
url. Defaults to "PHARMACODB".}

\item{port}{`character(1)` Name of enviromental variable with database
port. Defaults to "PHARMACODB_PORT".}

\item{user}{`character(1)` Name of environmental variable with database
username. Defaults to "MYSQL_USER".}

\item{password}{`character(1)` Name of environmental variable with database
user password. Defaults to "MYSQL_PW".}

\item{file}{`character(1)` Filepath of database credential file. Read
into the R environment to populate environmental variables for other
parameters. Defaults to file.path("~", ".mysql").}
}
\value{
`data.table` Table of tissue specific biomarkers from PharmacoDB.ca.
}
\description{
NOTE: this function requires valid PharmacoDB database
  credentials to work. Not intended for end users.
}
