% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetAnalysis.R
\name{getGeneSets}
\alias{getGeneSets}
\title{Get gene sets based on a query gene}
\usage{
getGeneSets(geneId, queryType)
}
\arguments{
\item{geneId}{The ENSEMBL ID of the gene you want to query}

\item{queryType}{The type of query you want to perform; see MSigDb for
possible subcategories (use msigdbr::msigdbr_collections())}
}
\value{
A data.table containing all the gene sets in the category queryType
and containing the query gene. data.table contains gene set ID, name, source,
URL, and description.
}
\description{
Get all gene sets from MSigDb that contain the query gene. Gene sets can be
based on biological process, cellular component, pathways, etc.
}
\examples{
getGeneSets("ENSG00000000971", "GO:BP")

}
