% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetAnalysis.R
\name{overlapDistance}
\alias{overlapDistance}
\title{Helper function for computeGeneSetSimilarity}
\usage{
overlapDistance(geneSets, gs1, gs2)
}
\arguments{
\item{geneSets}{A data.table with columns "gs_name" (gene set name) and
"ensembl_gene" (ENSEMBL gene ID) which lists the genes in each gene set}

\item{gs1}{The name of gene set 1}

\item{gs2}{The name of gene set 2}
}
\value{
The overlap distance, as defined in the description, between the two
gene sets
}
\description{
Compute the overlap distance between two gene sets, defined as (number of
overlapping genes) / (number of unique genes in the union of gene sets).
}
