% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA.R
\name{performSSGSEA}
\alias{performSSGSEA}
\title{Return a data.frame containing the results of single-sample gene set enrichment analyis
on a patient sample.}
\usage{
performSSGSEA(patientDf, geneSetsJSON)
}
\arguments{
\item{patientDf}{`data.frame` Single patient sample gene expression table, where 
rownames are genes and only one column is present for the expression of each gene}

\item{geneSetsJSON}{`character` This is the path to the JSON file for the gene set collection. 
The JSON file must contain two objects. A `name` which correspnods to a character
value containing the name of the gene set collection, and a `genesets` object which 
contains all the gene sets for that collection.}
}
\value{
`data.frame` ssGSEA output Data.Frame with rownames corresponding to gene sets 
and one column containing the ssGSEA estimates (between -1 to 1) for each gene set
}
\description{
Return a data.frame containing the results of single-sample gene set enrichment analyis
on a patient sample.
}
