% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfallPlot.R
\name{processXevaSet}
\alias{processXevaSet}
\title{Prepare a XevaSet for input into buildWaterfallPlot}
\usage{
processXevaSet(xevaSet, tissue, compound, mDataType)
}
\description{
Prepare a XevaSet for input into buildWaterfallPlot
}
\examples{
data(PDXE)
processXevaSet(PDXE, "BRCA", "paclitaxel", "RNASeq")

}
