% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetAnalysis.R
\name{queryGene}
\alias{queryGene}
\title{Query a gene}
\usage{
queryGene(geneId, queryType)
}
\arguments{
\item{geneId}{The ENSEMBL ID of the gene you want to query}

\item{queryType}{The type of query you want to perform; see MSigDb for
possible subcategories (use msigdbr::msigdbr_collections())}
}
\value{
A character vector containing the IDs of all gene sets this gene
is part of.
}
\description{
Get all gene sets from MSigDb that contain the query gene. Gene sets can be
based on biological process, cellular component, pathways, etc.
}
\examples{
queryGene("ENSG00000000971", "GO:BP")

}
