% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveDiffExpressedGenes.R
\name{retrieveDiffExpressedGenes}
\alias{retrieveDiffExpressedGenes}
\title{Return data frame of relevant biomarkers based on patient's differential gene expression
based on genes with expression >90th percentile and <10th percentile.}
\usage{
retrieveDiffExpressedGenes(patientDF, referenceDF, biomarkerDF)
}
\arguments{
\item{patientDF}{`data.frame` Patient sample gene expression table, where 
rownames are genes and only one column is present for the expression of each gene}

\item{referenceDF}{`data.frame` where rows are features (i.e. genes), columns
are samples, and values are a measurement on each feature for each sample.
Rownames must match rownames in `patientDF`}

\item{biomarkerDF}{`data.frame` This is the PharcoDB dataframe, containing all biomarkers
Every row is a biomarker association. Gene names in `patientDF` and}
}
\value{
`data.frame` Data frame containing rows with all biomarkers (from PharmacoDB)
relevant to the patient's overexpressed and underexpressed genes. Each row is a
biomarker association.
}
\description{
Return data frame of relevant biomarkers based on patient's differential gene expression
based on genes with expression >90th percentile and <10th percentile.
}
