% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPGxVision.R
\name{runPGxVision}
\alias{runPGxVision}
\title{Launch the shinydashboard for PGxVision}
\usage{
runPGxVision()
}
\value{
No return value; open up a shiny page
}
\description{
TODO: description
}
\examples{
\dontrun{
runPGxVision()
}

}
