% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TestGeneSets}
\alias{TestGeneSets}
\title{A set of 10 'GO cellular compartment' gene sets}
\format{
A data frame with 8187 rows and 2 variables:
\describe{
  \item{gs_id}{GO gene set ID}
  \item{ensembl_gene}{ENSEMBL gene ID}
}
}
\source{
Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B.
L., Gillette, M. A., … Mesirov, J. P. (2005). Gene set enrichment analysis:
A knowledge-based approach for interpreting genome-wide expression profiles.
_Proceedings of the National Academy of Sciences, 102_(43), 15545–15550.
doi:10.1073/pnas.0506580102
}
\usage{
TestGeneSets
}
\description{
A dataset of all gene sets in the category 'GO cellular compartment' that
contain the gene ENSG00000012124. Each row represents a single gene in one
of the 10 gene sets.
}
\keyword{datasets}
