% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA.PDXE.paxlitaxel.response}
\alias{BRCA.PDXE.paxlitaxel.response}
\title{Paxlitaxel response in BRCA PDXs}
\format{
A data frame with 38 rows and 3 variables:
\describe{
  \item{tumour}{The name of the PDX tumour}
  \item{ODC1}{The transcript levels of the ODC1 gene in the tumour}
  \item{angle}{The angle between the treatment and control}
}
}
\source{
Mer A, Haibe-Kains B (2021). Xeva: Analysis of patient-derived
xenograft (PDX) data. R package version 1.10.0.
}
\usage{
BRCA.PDXE.paxlitaxel.response
}
\description{
A dataset containing the response of different BRCA tumours to paxlitaxel,
as well as the levels of ODC1 transcripts in each tumour. The tumours are
patient-derived xenographs.
}
\keyword{datasets}
