% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{buildNetworkPlot}
\alias{buildNetworkPlot}
\title{Build a network plot of gene sets}
\usage{
buildNetworkPlot(gsSimilarityDf, similarityCutoff = 0.5, title = NULL)
}
\arguments{
\item{gsSimilarityDf}{A data.frame listing the similarity score for each
pair of gene sets. Must have columns "gs1", "gs2", "similarity"}

\item{similarityCutoff}{(optional) A number indicating the minimum
similarity two gene sets must have in order for an edge to show up on the
plot. Defaults to 0.5. The higher this value is, the more legible the plot
will be.}

\item{title}{(optional) A custom title for the network plot. Defaults to
"Gene Set Similarity Plot"}
}
\value{
invisible NULL
}
\description{
Create a network plot where each node corresponds to a gene set and edge
weights correspond to the similarity score between two gene sets.
}
\examples{
geneSetIds <- queryGene("ENSG00000000971", "GO:BP")
geneSets <- expandGeneSets(geneSetIds, "GO:BP")
gsSimilarity <- computeGeneSetSimilarity(geneSets)
buildNetworkPlot(gsSimilarityDf, similarityCutoff=0.3)

}
