% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{buildVolcanoPlot}
\alias{buildVolcanoPlot}
\title{Build a volcano plot of biomarkers}
\usage{
buildVolcanoPlot(biomarkerDf, experiment, pValCutoff = 0.01)
}
\arguments{
\item{biomarkerDf}{A data.frame of drug sensitivity biomarkers that you want
to plot, including the compound, tissue, and molecular data type used in the
experiment, the genomic coordinates of the genes which were tested on, and
the resulting p-value/fdr and estimate}

\item{experiment}{A named character vector representing the experiment for
which you want to plot biomarkers; an experiment is defined by a "tissue",
"compound", and "mDataType" (molecular data type)}

\item{pValCutoff}{A decimal number indicating the cutoff value for
significant observations; any results with a higher p-value will be grayed
out on the plot. Default value is 0.01.}
}
\value{
A ggplot2 plot object mapping the biomarkers of the experiment
(x-axis = estimate; y-axis = -log10(p-value))
}
\description{
Note that this volcano plot is NOT plotting results of differential gene
expression analysis but is plotting the drug sensitivity signature
}
\examples{
data(Biomarkers)
experiment <- c("Lung", "Trametinib", "rna")
names(experiment) <- c("tissue", "compound", "mDataType")
buildVolcanoPlot(Biomarkers, experiment, 0.005)

}
