% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPlot.R
\name{geneSetAnalysis}
\alias{geneSetAnalysis}
\title{Perform a gene set analysis on a gene}
\usage{
geneSetAnalysis(geneId, queryType, similarityCutoff = NULL, title = NULL)
}
\arguments{
\item{geneId}{The ENSEMBL ID of the gene you want to query}

\item{queryType}{The type of query you want to perform; see MSigDb for
possible subcategories (use msigdbr::msigdbr_collections())}

\item{similarityCutoff}{(optional) A number indicating the minimum
similarity two gene sets must have in order for an edge to show up on the
plot. Defaults to 0.5.}

\item{title}{(optional) A custom title for the network plot. Defaults to
"querytype Gene Sets containing geneId (edge weights based on gene set
overlap)".}
}
\value{
The igraph object containing the network plot of gene sets
}
\description{
Query MSigDb to find all gene sets that queryGene is in, then compute the
similarity of the gene sets and plot them on a network graph. This is a
wrapper function which performs the full gene set analysis pipeline and
plots the result.
}
\examples{
geneSetAnalysis("ENSG00000000971", "GO:BP", 0.3)

}
