% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetAnalysis.R
\name{geneSetAnalysis}
\alias{geneSetAnalysis}
\title{Perform a gene set analysis on a gene}
\usage{
geneSetAnalysis(geneId, queryType, similarityMetric = "overlap")
}
\arguments{
\item{geneId}{The ENSEMBL ID of the gene you want to query}

\item{queryType}{The type of query you want to perform; see MSigDb for
possible subcategories (use msigdbr::msigdbr_collections())}

\item{similarityMetric}{The algorithm used to compute the similarity between
gene sets. The default (and only option currently) is "overlap".}
}
\value{
TODO
}
\description{
Query MSigDb to find all gene sets that queryGene is in, then compute the
similarity of the gene sets and return the result in a data frame. This is a
wrapper function which performs the full gene set analysis pipeline, minus
the network plot.
}
\examples{
geneSetAnalysis("ENSG00000000971", "GO:BP")

}
