% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCh38.p13.Assembly}
\alias{GRCh38.p13.Assembly}
\title{GRCh38.p13 genome assembly}
\format{
A data frame with 25 rows and 3 variables:
\describe{
  \item{chrName}{the name of the chromosome (all, 1-22, X, Y)}
  \item{molecule-type/loc}{the molecule type ("chromosome" or "all")}
  \item{chrLength}{the length of the chromosome (# nucleotides)}
}
}
\source{
Frankish, A., Diekhans, M., Ferreira, A. M., Johnson, R., Jungreis,
I. Loveland, J., Mudge, J. M., Sisu, C., Wright, J., Armstrong, J., Barnes,
I., Berry, A., Bignell, A., Carbonell Sala, S., Chrast, J., Cunningham, F.,
Di Domenico, T., Donaldson, S., Fiddes, I. T., García Girón, C., … Flicek, P.
(2019). GENCODE reference annotation for the human and mouse genomes.
_Nucleic acids research, 47_(D1), D766–D773.
https://doi.org/10.1093/nar/gky955
}
\usage{
GRCh38.p13.Assembly
}
\description{
A dataset containing the characteristics of of the GRCh38.p13 genome
assembly, specifically chromosome names and lengths.
}
\keyword{datasets}
