% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCh38.p13.Assembly}
\alias{GRCh38.p13.Assembly}
\title{GRCh38.p13 genome assembly}
\format{
A data frame with 25 rows and 3 variables:
\describe{
  \item{chrName}{the name of the chromosome (all, 1-22, X, Y)}
  \item{molecule-type/loc}{the molecule type ("chromosome" or "all")}
  \item{chrLength}{the length of the chromosome (# nucleotides)}
}
}
\source{
\url{https://www.gencodegenes.org/pages/gencode.html}
}
\usage{
GRCh38.p13.Assembly
}
\description{
A dataset containing the characteristics of of the GRCh38.p13 genome
assembly, specifically chromosome names and lengths.
}
\keyword{datasets}
