% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfallPlot.R
\name{processXevaSet}
\alias{processXevaSet}
\title{Prepare a XevaSet for input into buildWaterfallPlot
tissue <- "BRCA"
mDataType <- "RNASeq"
compound <- "paclitaxel"}
\usage{
processXevaSet(xevaSet, tissue, compound, mDatatype)
}
\description{
Prepare a XevaSet for input into buildWaterfallPlot
tissue <- "BRCA"
mDataType <- "RNASeq"
compound <- "paclitaxel"
}
